Restructured project files to follow MVC pattern
