+++ src/main/java/cs/oswego/edu/controller/RestApplication.java
@@ -1,5 +1,3 @@
-package edu.oswego.edu.rest.Application;
-
 import javax.ws.rs.ApplicationPath;
 import javax.ws.rs.core.Application;
