--- src/main/java/edu/oswego/edu/rest/datastore/TeamNameDatastore.java
@@ -1,9 +1,0 @@
-package edu.oswego.edu.rest.datastore;
-
-import java.util.ArrayList;
-
-public class TeamNameDatastore {
-
-    public static ArrayList<String> teamName = new ArrayList<>();
-
-}
