+++ src/main/webapp/WEB-INF/package.json
