+++ src/main/webapp/WEB-INF/views/main.jsx
