+++ src/main/webapp/WEB-INF/vite.config.js
